Change MaceEvolve.Core to use 'float' instead of 'double'

+++ MaceEvolve.Core/Interfaces/IFood.cs
@@ -4,6 +4,6 @@
     {
         int Servings { get; set; }
         int EnergyPerServing { get; set; }
-        double ServingDigestionCost { get; set; }
+        float ServingDigestionCost { get; set; }
     }
 }
